requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pymongo==4.8.0
motor==3.5.1       
pydantic==2.9.2
python-jose==3.3.0  
passlib[bcrypt]==1.7.4  
python-dotenv==1.0.1
stripe==10.7.0     
qrcode==7.4.2     